fixed index insertion

+++ db/global_index.cc
@@ -7,7 +7,7 @@ GlobalIndex::GlobalIndex() {
 }
 
 const DataMeta* GlobalIndex::Get(const std::string& key) {
-  size_t hash = std::hash<std::string>{}(key);
+  uint32_t hash = (uint32_t) std::hash<std::string>{}(key);
   void *p = tree_->search(hash);
   return (const DataMeta *) p;
 }
@@ -19,7 +19,7 @@ void GlobalIndex::Add(const std::string& key, const uint64_t& offset, const uint
   meta->size = size;
   meta->file_meta = file_meta;
   clflush((char *) meta, sizeof(DataMeta));
-  int64_t hash = std::hash<std::string>{}(key);
+  uint32_t hash = (uint32_t) std::hash<std::string>{}(key);
   tree_->insert(hash, meta);
 }
 
@@ -30,7 +30,7 @@ void GlobalIndex::Update(const std::string& key, const uint64_t& offset, const u
   meta->size = size;
   meta->file_meta = file_meta;
   clflush((char *) meta, sizeof(DataMeta));
-  int64_t hash = std::hash<std::string>{}(key);
+  uint32_t hash = (uint32_t) std::hash<std::string>{}(key);
   tree_->update(hash, meta);
 }
